pyproject.toml
[project]
name = "resume-tailor"
version = "0.1.0"
description = "LangGraph-powered resume tailoring pipeline"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
parsers = [
    "pdfplumber>=0.10",
    "python-docx>=1.0",
]
research = [
    "langchain-community>=0.3.0",
    "tavily-python>=0.3",
]

[project.scripts]
resume-tailor = "main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
